requirements.txt
langchain-core
langchain-openai
langchain-anthropic
langchain-deepseek
langchain-google-genai
langgraph
langchain
langchain-text-splitters
openai
anthropic
pandas
scikit-learn
numpy
pyyaml
paramiko
ipython
pydantic
typing-extensions
dotenv